Move ExecutionContext to own file
