+++ src/Stratis.SmartContracts.CLR/StateProcessor.cs
@@ -5,18 +5,6 @@ using Stratis.SmartContracts.RuntimeObserver;
 
 namespace Stratis.SmartContracts.CLR
 {
-    public class ExecutionContext
-    {
-        public ExecutionContext(Observer observer)
-        {
-            this.Observer = observer;
-        }
-
-        public Observer Observer { get; }
-
-        public IGasMeter GasMeter => this.Observer.GasMeter;
-    }
-
     public class StateProcessor : IStateProcessor
     {
         public StateProcessor(IVirtualMachine vm,
